Fix fencepost error in gd_entry_list reported by Christian Trippe.  Also deal with (perlmandir + --with-perl-dir) and fix descriptor leak in flac encoding.

+++ test/lzma_put.c
@@ -47,7 +47,7 @@ int main(void)
   close(fd);
 
   fd = open(data, O_CREAT | O_EXCL | O_WRONLY | O_BINARY, 0666);
-  write(fd, data_data, 256 * sizeof(uint16_t));
+  write(fd, data_data, 256 * sizeof(uint8_t));
   close(fd);
 
   /* compress */
